twilio-voice-frontend/requirements.txt
twilio 
azure-cognitiveservices-speech 
Wave 
Flask 
flask-sock 
soundfile 
pyngrok
python-dotenv
langchain==0.1.1
ollama